Upgrade to maxminddb-golang/v2@v2.0.0-beta.7
In particular, this replaces the old deserializer interface with the
new Unmarshaler one.

+++ tree_test.go
@@ -104,7 +104,7 @@ func TestTreeInsertAndGet(t *testing.T) {
 		},
 		"double":      42.123456,
 		"float":       float32(1.1),
-		"int32":       -268435456,
+		"int32":       int32(-268435456),
 		"map":         allTypesLookupSubmap,
 		"uint128":     &bigInt,
 		"uint16":      uint64(0x64),
